stuff works but still syncing errors

+++ Assignment 5/Q1/defs.h
@@ -42,7 +42,7 @@ typedef struct student {
     int64_t pref1;
     int64_t pref2;
     int64_t pref3;
-    int64_t calibre;
+    float calibre;
     int64_t time;
 } student;
